requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
